ai/rag-engine/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_rag"
version = "0.1.0"
description = "Local RAG engine with ChromaDB and Ollama"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb~=0.5.3",
    "fastapi~=0.115.0",
    "uvicorn[standard]~=0.30.0",
    "httpx~=0.27.0",
    "pyyaml~=6.0.2",
    "sentence-transformers~=3.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
